FIX: Pipeline manager so it would run VFS-calls
Add: Registering calls for VFS functionality;

 - A normal Event Loop is implemented, which processes syscall signals (intercepted calls) and returns the result via syscall_return;

 - The logic for working with /dev/ has been fixed: PM doesn't know about g_tSymbolicLinks (this is a variable inside DKMS), so it should simply forward the path as is, and let DKMS figure it out itself;

 - Fixed parsing of variables in the main while true loop;

 - Fixed dkms_complete_irp function to accept and record error status and added check for pIrp existence before accessing it (crash protection);

----------------------

Note: If DKMS is waiting for a driver to initialize and another signal comes in (for example, a new boot request), it should save that signal for later rather than consider it an error. The code can now 'defer' signals while waiting for the driver to load. For the next commit, keep in mind that the kernel in our current form adds true to the beginning when resuming the process after yield (this happens when calling intercepted syscalls of the dkms_* type), but does not add it for regular syscalls (of the raw_* type).